ig/requirements.txt
groq>=1.2.0
instagrapi>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
pytz>=2024.0
playwright>=1.40.0
requests>=2.31.0
